refactor: @web-widget/builder -> @web-widget/vite
