+++ examples/examples/package.json
@@ -37,7 +37,7 @@
     "@types/react-dom": "^18.0.6",
     "@vitejs/plugin-react": "^4.0.0",
     "@vitejs/plugin-vue": "^4.2.3",
-    "@web-widget/builder": "workspace:*",
+    "@web-widget/vite": "workspace:*",
     "build-worker": "^0.2.2",
     "typescript": "^5.0.4",
     "vite": "^4.4.9"
