+++ examples/examples/vite.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
 import react from "@vitejs/plugin-react";
-import { webRouterPlugin } from "@web-widget/builder";
+import { webRouterPlugin } from "@web-widget/vite";
 import reactWebWidgetPlugin from "@web-widget/react/vite";
 import vueWebWidgetPlugin from "@web-widget/vue/vite";
