+++ examples/react-and-vue2/package.json
@@ -35,7 +35,7 @@
     "@types/react-dom": "^18.0.6",
     "@vitejs/plugin-react": "^4.0.0",
     "@vitejs/plugin-vue2": "^2.2.0",
-    "@web-widget/builder": "workspace:*",
+    "@web-widget/vite": "workspace:*",
     "tsup": "^6.7.0",
     "typescript": "^5.0.4",
     "vite": "^4.4.9"
