+++ packages/react/package.json
@@ -37,7 +37,7 @@
     "prepublishOnly": "pnpm run build"
   },
   "dependencies": {
-    "@web-widget/builder": "workspace:*",
+    "@web-widget/vite": "workspace:*",
     "@web-widget/schema": "workspace:*",
     "@web-widget/web-widget": "workspace:*"
   },
