+++ packages/vite/src/config.ts
