+++ packages/vite/src/index.ts
