+++ packages/vite/src/server/loader/index.ts
