+++ packages/vite/src/server/loader/loader.ts
