+++ packages/vite/src/server/loader/vite.ts
