+++ packages/vite/src/server/render/util.ts
