+++ packages/vite/src/web-widget.ts
