+++ packages/vite/tsconfig.json
