+++ packages/vite/tsup.config.ts
@@ -1,7 +1,7 @@
 import type { Options } from "tsup";
 export const tsup: Options = {
   entry: {
-    builder: "src/index.ts",
+    vite: "src/index.ts",
   },
   dts: true,
   target: "node18",
