+++ packages/vue/package.json
@@ -35,7 +35,7 @@
     "prepublishOnly": "pnpm run build"
   },
   "dependencies": {
-    "@web-widget/builder": "workspace:*",
+    "@web-widget/vite": "workspace:*",
     "@web-widget/html": "workspace:*",
     "@web-widget/schema": "workspace:*",
     "@web-widget/web-widget": "workspace:*"
