+++ packages/vue2/package.json
@@ -30,7 +30,7 @@
     "prepublishOnly": "pnpm run build"
   },
   "dependencies": {
-    "@web-widget/builder": "workspace:*",
+    "@web-widget/vite": "workspace:*",
     "@web-widget/schema": "workspace:*",
     "@web-widget/web-widget": "workspace:*",
     "vue": "^2.7.14",
